+++ test/invariant/DivergenceTrade.invarant.t.sol
@@ -6,14 +6,14 @@ import { DeploymentFixture } from "../fixtures/Ready.sol";
 import { TestERC20 } from "../shared/TestERC20.sol";
 import { Handler } from "../invariant/handlers/Handler.sol";
 import { IERC721Enumerable, IERC721 } from "@oz/token/ERC721/extensions/IERC721Enumerable.sol";
-import { LiquidityType, Outcome } from "../../src/core/types/common.sol";
+import { LiquidityType, Outcome } from "core/types/common.sol";
 import { IManagerState } from "../../src/periphery/interfaces/IManagerState.sol";
 import { IERC20 } from "@oz/token/ERC20/IERC20.sol";
-import { IBattleState } from "../../src/core/interfaces/battle/IBattleState.sol";
+import { IBattleState } from "core/interfaces/battle/IBattleState.sol";
 import { console2 } from "@std/console2.sol";
 import { Multicall } from "@oz/utils/Multicall.sol";
 import { Position, PositionState } from "../../src/periphery/types/common.sol";
-import { IBattleTrade } from "../../src/core/interfaces/battle/IBattleActions.sol";
+import { IBattleTrade } from "core/interfaces/battle/IBattleActions.sol";
 import { IQuoter } from "../../src/periphery/interfaces/IQuoter.sol";
 
 // contract DivergenceInvariant is DeploymentFixture {
@@ -65,13 +65,12 @@ contract DivergenceTradeInvariant is Test {
         // uint256 totalSpear = IERC20(handler.spear()).totalSupply();
         // uint256 totalShield = IERC20(handler.shield()).totalSupply();
         // It will check something to prove trades are ok.
-        uint cAmount = handler.ghost_total_camount();
-        uint spearAmount = handler.ghost_total_samount_spear();
-        uint shieldAmount = handler.ghost_total_samount_shield();
-        uint col = handler.ghost_seed_collateral();
-        uint sAmount = spearAmount > shieldAmount ? spearAmount : shieldAmount;
-        assertGe(cAmount+col, sAmount);
+        uint256 cAmount = handler.ghost_total_camount();
+        uint256 spearAmount = handler.ghost_total_samount_spear();
+        uint256 shieldAmount = handler.ghost_total_samount_shield();
+        uint256 col = handler.ghost_seed_collateral();
+        uint256 sAmount = spearAmount > shieldAmount ? spearAmount : shieldAmount;
+        assertGe(cAmount + col, sAmount);
         console2.log("**************************");
     }
-
 }
